--- cvInterApp/package.json
@@ -1,12 +1,0 @@
-{
-  "devDependencies": {
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@types/react-router-dom": "^5.3.3",
-    "@vitejs/plugin-react": "^4.3.1",
-    "vite": "^5.3.4"
-  },
-  "dependencies": {
-    "react-router-dom": "^6.25.1"
-  }
-}
