--- cvInterApp/src/application/recruiter/Components/Recruiter.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-import CandidateSearch from "./CandidateSearch";
-import SideBar from "./SideBar";
-
-
-
-const Recruiter = () => {
-  return (
-    <div>
-        <SideBar />
-        <CandidateSearch />
-      
-    </div>
-  );
-};
-
-export default Recruiter
